Started refactoring floating point value constructors, WIP
