+++ lib/src/arithmetic/floating_point/floating_point.dart
@@ -4,4 +4,4 @@
 export 'floating_point_adder_round.dart';
 export 'floating_point_adder_simple.dart';
 export 'floating_point_logic.dart';
-export 'floating_point_value.dart';
+export 'floating_point_values/floating_point_values.dart';
